Stop deleting fragments after compiling.
This required a new verison directory structure, where unversioned
fragments are stored in `next` and fragments are moved from `next` when
a changelog is compiled for a particular version.

Additionally side effects have been refactored and isolated. This makes
it possible to implement dry run with a reasonable degree of confidence
that it won't make any changes.